Update Create Bookmark from Current Tab in Chrome.applescript
Added ability to populate Spillo's Description field with the currently selected text in the Chrome tab. (It is optional behavior: if no selection, Description is not populated.) Note: Due to Chrome's refusal to execute javascript via Applescript, we must rely on the clipboard, so the current clipboard contents are replaced with selected text when this script runs.